requirements.txt
rhasspy-asr==0.1.5
google-cloud-speech==1.3.1
